+++ tests/__init__.py
@@ -1,0 +1,7 @@
+"""
+Micro_Tracker测试套件
+
+Author: Lucien (lucien-6@qq.com)
+Date: 2025-11-01
+"""
+